added account page

+++ AskAHuman/Services/Interfaces/IUserService.cs
@@ -8,7 +8,7 @@ namespace AskAHuman.Services.Interfaces;
 public interface IUserService
 {
     public List<UserDTO> GetAll();
-    public Result<User> GetById(int id);
+    public Result<User> GetById(long id);
     public string? Login(string username, string password);
     public bool Register(string username, string password);
     public bool UserExists(string username);
